updates based on reviews
